Throw an exception when not decoding
When trying to decode a file that is not a png, `decode` will remain in an infinite loop due to `while(true)` and the fact that  `var len = bin.readUint(data, offset)` gives `NaN`. Then, the method continues to read the file, adding `4` to the offset but actually adding 4 to `NaN`, which gives `NaN`, and so on.
The quick fix is to throw an exception whenever the `len` is `NaN` because we now it would then be trapped.